# Add variants capacity