… sample-playlists/media-playlist-large.m3u8