* Add variants capacity of MasterPlaylist

Before:
```
BenchmarkDecodeMasterPlaylist-8	   30000	     46001 ns/op	   12197 B/op	      77 allocs/op
BenchmarkDecodeMediaPlaylist-8 	     100	  17027700 ns/op	11000965 B/op	  120105 allocs/op
BenchmarkEncodeMasterPlaylist-8	 1000000	      1066 ns/op	      40 B/op	      11 allocs/op
BenchmarkEncodeMediaPlaylist-8 	     300	   4794475 ns/op	   45424 B/op	     405 allocs/op
```

After:
```
BenchmarkDecodeMasterPlaylist-8	   30000	     45144 ns/op	   12205 B/op	      74 allocs/op
BenchmarkDecodeMediaPlaylist-8 	     100	  17042414 ns/op	11000942 B/op	  120102 allocs/op
BenchmarkEncodeMasterPlaylist-8	 1000000	      1062 ns/op	      40 B/op	      11 allocs/op
BenchmarkEncodeMediaPlaylist-8 	     300	   4703589 ns/op	   45426 B/op	     405 allocs/op
```

Benchstat Summary
```
name                    old time/op    new time/op    delta
DecodeMasterPlaylist-8    46.0µs ± 0%    45.1µs ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     17.0ms ± 0%    17.0ms ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8    1.07µs ± 0%    1.06µs ± 0%   ~             (p=1.000 n=1+1)
EncodeMediaPlaylist-8     4.79ms ± 0%    4.70ms ± 0%   ~             (p=1.000 n=1+1)

name                    old alloc/op   new alloc/op   delta
DecodeMasterPlaylist-8    12.2kB ± 0%    12.2kB ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8     11.0MB ± 0%    11.0MB ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8     40.0B ± 0%     40.0B ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8     45.4kB ± 0%    45.4kB ± 0%   ~             (p=1.000 n=1+1)

name                    old allocs/op  new allocs/op  delta
DecodeMasterPlaylist-8      77.0 ± 0%      74.0 ± 0%   ~             (p=1.000 n=1+1)
DecodeMediaPlaylist-8       120k ± 0%      120k ± 0%   ~             (p=1.000 n=1+1)
EncodeMasterPlaylist-8      11.0 ± 0%      11.0 ± 0%   ~     (all samples are equal)
EncodeMediaPlaylist-8        405 ± 0%       405 ± 0%   ~     (all samples are equal)
```