@bradleyfalzon 2016-07-06
Hi,

I'm not sure whether we really want to make this change, as the time spend in a Master Manifest is trivial, and reducing the allocations isn't a substantial cost, this can also be achieved in your client, eg:

```
master := m3u8.NewMasterPlaylist()
master.Variants = make([]*Variant, 0, 16)
```