@hori-ryota 2016-07-06
That’s true.